backend/response_generation/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
assemblyai==0.35.1
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
Flask==3.1.0
Flask-Sockets==0.2.1
frozenlist==1.5.0
gevent==24.11.1
gevent-websocket==0.10.1
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
pip3-autoremove==1.2.2
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
PyJWT==2.10.0
pyparsing==3.2.0
python-dotenv==1.0.1
requests==2.32.3
rev_ai==2.20.0
rsa==4.9
scipy==1.14.1
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
tqdm==4.67.0
twilio==9.3.7
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
websocket-client==0.59.0
websockets==14.1
Werkzeug==3.1.3
yarl==1.18.0
zope.event==5.0
zope.interface==7.1.1